Implemented DTO layer for Ticket,TicketComment,Campaign modules

+++ src/main/java/com/venkatesh/ai_crm_platform/Service/CampaignService.java
@@ -1,9 +1,13 @@
 package com.venkatesh.ai_crm_platform.Service;
 
-
 import com.venkatesh.ai_crm_platform.Repository.CampaignRepository;
+import com.venkatesh.ai_crm_platform.Repository.UserRepository;
+import com.venkatesh.ai_crm_platform.dto.campaign.CampaignRequestDto;
+import com.venkatesh.ai_crm_platform.dto.campaign.CampaignResponseDto;
 import com.venkatesh.ai_crm_platform.exception.ResourceNotFoundException;
+import com.venkatesh.ai_crm_platform.mapper.CampaignMapper;
 import com.venkatesh.ai_crm_platform.models.Entities.Campaign;
+import com.venkatesh.ai_crm_platform.models.Entities.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,41 +18,67 @@ import java.util.List;
 public class CampaignService {
 
     private final CampaignRepository campaignRepository;
+    private final UserRepository userRepository;
 
-    public Campaign create(Campaign campaign){
-        return campaignRepository.save(campaign);
-    }
+    public CampaignResponseDto create(CampaignRequestDto request){
 
-    public List<Campaign> getAll(){
-        return campaignRepository.findAll();
-    }
+        Campaign campaign = CampaignMapper.toEntity(request);
 
-    public Campaign getById(Long id){
-        return campaignRepository.findById(id)
+        User user = userRepository.findById(request.getCreatedById())
                 .orElseThrow(() ->
-                        new ResourceNotFoundException(
-                                "Campaign Not Found"));
+                        new ResourceNotFoundException("User Not Found"));
+
+        campaign.setCreatedBy(user);
+
+        return CampaignMapper.toResponse(
+                campaignRepository.save(campaign));
     }
 
-    public Campaign update(Long id,
-                           Campaign campaign){
+    public List<CampaignResponseDto> getAll(){
 
-        Campaign existing = getById(id);
+        return campaignRepository.findAll()
+                .stream()
+                .map(CampaignMapper::toResponse)
+                .toList();
+    }
 
-        existing.setName(campaign.getName());
-        existing.setDescription(
-                campaign.getDescription());
+    public CampaignResponseDto getById(Long id){
 
-        existing.setStartDate(
-                campaign.getStartDate());
+        Campaign campaign = campaignRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Campaign Not Found"));
 
-        existing.setEndDate(
-                campaign.getEndDate());
+        return CampaignMapper.toResponse(campaign);
+    }
 
-        return campaignRepository.save(existing);
+    public CampaignResponseDto update(Long id,
+                                      CampaignRequestDto request){
+
+        Campaign campaign = campaignRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Campaign Not Found"));
+
+        campaign.setName(request.getName());
+        campaign.setDescription(request.getDescription());
+        campaign.setStartDate(request.getStartDate());
+        campaign.setEndDate(request.getEndDate());
+
+        User user = userRepository.findById(request.getCreatedById())
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User Not Found"));
+
+        campaign.setCreatedBy(user);
+
+        return CampaignMapper.toResponse(
+                campaignRepository.save(campaign));
     }
 
     public void delete(Long id){
+
+        if(!campaignRepository.existsById(id)){
+            throw new ResourceNotFoundException("Campaign Not Found");
+        }
+
         campaignRepository.deleteById(id);
     }
 }
