+++ src/main/java/com/venkatesh/ai_crm_platform/Service/TicketCommentService.java
@@ -1,11 +1,19 @@
 package com.venkatesh.ai_crm_platform.Service;
 
 import com.venkatesh.ai_crm_platform.Repository.TicketCommentRepository;
+import com.venkatesh.ai_crm_platform.Repository.TicketRepository;
+import com.venkatesh.ai_crm_platform.Repository.UserRepository;
+import com.venkatesh.ai_crm_platform.dto.ticketcomment.TicketCommentRequestDto;
+import com.venkatesh.ai_crm_platform.dto.ticketcomment.TicketCommentResponseDto;
 import com.venkatesh.ai_crm_platform.exception.ResourceNotFoundException;
+import com.venkatesh.ai_crm_platform.mapper.TicketCommentMapper;
+import com.venkatesh.ai_crm_platform.models.Entities.Ticket;
 import com.venkatesh.ai_crm_platform.models.Entities.TicketComment;
+import com.venkatesh.ai_crm_platform.models.Entities.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -13,23 +21,54 @@ import java.util.List;
 public class TicketCommentService {
 
     private final TicketCommentRepository ticketCommentRepository;
+    private final TicketRepository ticketRepository;
+    private final UserRepository userRepository;
 
-    public TicketComment create(TicketComment comment){
-        return ticketCommentRepository.save(comment);
-    }
+    public TicketCommentResponseDto create(TicketCommentRequestDto request){
 
-    public List<TicketComment> getAll(){
-        return ticketCommentRepository.findAll();
-    }
+        TicketComment comment =
+                TicketCommentMapper.toEntity(request);
 
-    public TicketComment getById(Long id){
-        return ticketCommentRepository.findById(id)
+        comment.setCreatedAt(LocalDateTime.now());
+
+        Ticket ticket = ticketRepository.findById(request.getTicketId())
                 .orElseThrow(() ->
-                        new ResourceNotFoundException(
-                                "TicketComment Not Found"));
+                        new ResourceNotFoundException("Ticket Not Found"));
+
+        User user = userRepository.findById(request.getUserId())
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User Not Found"));
+
+        comment.setTicket(ticket);
+        comment.setUser(user);
+
+        return TicketCommentMapper.toResponse(
+                ticketCommentRepository.save(comment));
+    }
+
+    public List<TicketCommentResponseDto> getAll(){
+
+        return ticketCommentRepository.findAll()
+                .stream()
+                .map(TicketCommentMapper::toResponse)
+                .toList();
+    }
+
+    public TicketCommentResponseDto getById(Long id){
+
+        TicketComment comment = ticketCommentRepository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Comment Not Found"));
+
+        return TicketCommentMapper.toResponse(comment);
     }
 
     public void delete(Long id){
+
+        if(!ticketCommentRepository.existsById(id)){
+            throw new ResourceNotFoundException("Comment Not Found");
+        }
+
         ticketCommentRepository.deleteById(id);
     }
 }
